requirements.txt

Django>=4.2,<5.0
Pillow>=10.0
gunicorn
